+++ pvDataJava/documentation/RELEASE_NOTES.md
@@ -3,11 +3,21 @@ Release release/3.1 IN DEVELOPMENT
 
 The main changes since release 3.0.2 are:
 
+* methods that change interface removed from PVField and PVStructure
 * union is new type.
 * copy is new.
 * monitorPlugin is new.
 * PVField no longer extends Requester
 
+methods removed from  PVField and PVStructure
+-----------------
+
+The following method was removed from  PVField: rename.
+The following methods were removed from PVStrucure: appendPVField, appendPVFields, removePVField,
+replacePVField, getExtendsStructureName, and putExtendsStructureName.
+
+With these changes there should be no methods that can change the introspection interface
+of any data onject after it is created.
 
 union is a new basic type.
 ------------